Struct glyph_brush_draw_cache::DrawCache

pub struct DrawCache { /* private fields */ }

Dynamic rasterization draw cache.

Implementations

impl DrawCache

pub fn builder() -> DrawCacheBuilder

Returns a default DrawCacheBuilder.

pub fn scale_tolerance(&self) -> f32

Returns the current scale tolerance for the cache.

pub fn position_tolerance(&self) -> f32

Returns the current subpixel position tolerance for the cache.

pub fn dimensions(&self) -> (u32, u32)

Returns the cache texture dimensions assumed by the cache. For proper operation this should match the dimensions of the used GPU texture.

pub fn queue_glyph(&mut self, font_id: usize, glyph: Glyph)

Queue a glyph for caching by the next call to cache_queued. font_id is used to disambiguate glyphs from different fonts. The user should ensure that font_id is unique to the font the glyph is from.

pub fn clear(&mut self)

Clears the cache. Does not affect the glyph queue.

pub fn clear_queue(&mut self)

Clears the glyph queue.

pub fn to_builder(&self) -> DrawCacheBuilder

Returns a DrawCacheBuilder with this cache’s attributes.

pub fn cache_queued<F, U>( &mut self, fonts: &[F], uploader: U ) -> Result<CachedBy, CacheWriteErr>where F: Font + Sync, U: FnMut(Rectangle<u32>, &[u8]),

Caches the queued glyphs. If this is unsuccessful, the queue is untouched. Any glyphs cached by previous calls to this function may be removed from the cache to make room for the newly queued glyphs. Thus if you want to ensure that a glyph is in the cache, the most recently cached queue must have contained that glyph.

uploader is the user-provided function that should perform the texture uploads to the GPU. The information provided is the rectangular region to insert the pixel data into, and the pixel data itself. This data is provided in horizontal scanline format (row major), with stride equal to the rectangle width.

If successful returns a CachedBy that can indicate the validity of previously cached glyph textures.

pub fn rect_for(&self, font_id: usize, glyph: &Glyph) -> Option<TextureCoords>

Retrieves the (floating point) texture coordinates of the quad for a glyph in the cache, as well as the pixel-space (integer) coordinates that this region should be drawn at. These pixel-space coordinates assume an origin at the top left of the quad. In the majority of cases these pixel-space coordinates should be identical to the bounding box of the input glyph. They only differ if the cache has returned a substitute glyph that is deemed close enough to the requested glyph as specified by the cache tolerance parameters.

A sucessful result is Some if the glyph is not an empty glyph (no shape, and thus no rect to return).

Ensure that font_id matches the font_id that was passed to queue_glyph with this glyph.